+++ lib/rustyjson.ex
@@ -365,7 +365,8 @@ defmodule RustyJson do
       "x86_64-unknown-linux-gnu" => x86_64_variants,
       "x86_64-apple-darwin" => x86_64_variants,
       "x86_64-pc-windows-msvc" => x86_64_variants,
-      "x86_64-pc-windows-gnu" => x86_64_variants
+      "x86_64-pc-windows-gnu" => x86_64_variants,
+      "x86_64-unknown-linux-musl" => x86_64_variants
     },
     version: version